+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `roslyn5.0` NuGet package flavor (`analyzers/dotnet/roslyn5.0/cs`) ([PR](https://github.com/dotnet/roslynator/pull/1787))
+- Add analyzer "Call 'IsKind' instead of comparing 'Kind'" ([RCS9012](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS9012)) ([PR](https://github.com/dotnet/roslynator/pull/1820))
 
 ### Breaking